backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pillow==10.2.0
rembg==2.0.50
numpy==1.26.3
opencv-python==4.9.0.80
requests==2.31.0
pydantic-settings==2.1.0
redis==5.0.1
